requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
fastapi==0.95.0
uvicorn==0.22.0